完成了拦截器的编写和注册，实现了添加员工功能

+++ sky-pojo/src/main/java/entity/Employee.java
@@ -23,6 +23,6 @@ public class Employee implements Serializable {
     private Integer status;
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
-    private Integer createUser;
-    private Integer updateUser;
+    private Long createUser;
+    private Long updateUser;
 }
